backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
radon==6.0.1
gitpython==3.1.43
pydantic==2.9.0
python-multipart==0.0.9
fpdf2==2.8.1
aiofiles==24.1.0
